backend/requirements.txt
Django>=5.2.0
djangorestframework>=3.16.0
djangorestframework-simplejwt>=5.5.0
python-decouple>=3.8
drf-spectacular>=0.29.0
psycopg2-binary>=2.9.0
Pillow>=12.0.0
dj-database-url>=2.1.0
django-cors-headers>=4.3.0
gunicorn>=21.2.0
whitenoise>=6.6.0